# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Stock Management API",
        "description": "REST API untuk mengelola data inventaris, produk, supplier, warehouse, dan stok.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get All Suppliers",
                "operationId": "6207e57dd1dcd1d6e7125fe144607196",
                "responses": {
                    "200": {
                        "description": "List of suppliers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create Supplier",
                "operationId": "dc5dc7ccb140cd5b81cab6f42b396691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT Dahana"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Subang, Jawa Barat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully"
                    }
                }
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get Supplier Detail",
                "operationId": "7fc447855461518ce36bdeb5a44310a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update Supplier",
                "operationId": "22145931fdda85497cf90c73c5406747",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT Dahana Indonesia"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Bandung"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete Supplier",
                "operationId": "2fe04361f0f1581a9f5dbb4094b634b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier deleted successfully"
                    }
                }
            }
        },
        "/api/warehouses": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get All Warehouses",
                "operationId": "0c48cbbd5731b5e56eda7c180576670b",
                "responses": {
                    "200": {
                        "description": "List of warehouses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Create Warehouse",
                "operationId": "c1b0005f758b5892517cade2be77a1d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Warehouse A"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Subang"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Main Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Warehouse created successfully"
                    }
                }
            }
        },
        "/api/warehouses/{id}": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get Warehouse Detail",
                "operationId": "9f201c10112ab3904ca3aa4e71e63e89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Update Warehouse",
                "operationId": "8d7614858df133388092bda00a4be309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Warehouse B"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warehouse updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Delete Warehouse",
                "operationId": "45d106d3e2367cdc02fc89c04fc702c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse deleted successfully"
                    }
                }
            }
        }
    },
...
}